+++ cocos/core/utils/internal.ts
